Reapply "Font size: em instead of pixels, rem-em-px conversion fix (closes #1171)"
This reverts commit e3f5d02b041a25ef94e0d02e400e35b98fe40eeb.

+++ web-client/src/main/java/lsfusion/gwt/client/base/size/GSize.java
@@ -31,10 +31,10 @@ public abstract class GSize implements Serializable {
     }
 
     /* the trickiest part we have the exact size but we want the relative size */
-    public static GSize getCalcValueSize(int pixels, int fontSize) { // "font dependent" pixels (not explicit constants like in getValueSize)
+    public static GSize getCalcValueSize(int pixels, double fontSize) { // "font dependent" pixels (not explicit constants like in getValueSize)
         return GFixedSize.getSize(pixels, VALUE_TYPE, fontSize);
     }
-    public static GSize getCalcComponentSize(int pixels, int remFontSize) { // "font dependent" pixels (not explicit constants like in getValueSize)
+    public static GSize getCalcComponentSize(int pixels, double remFontSize) { // "font dependent" pixels (not explicit constants like in getValueSize)
         return GFixedSize.getSize(pixels, COMPONENT_TYPE, remFontSize);
     }
